+++ controller/terminatinggateway_controller.go
@@ -1,0 +1,42 @@
+package controller
+
+import (
+	"context"
+
+	"github.com/go-logr/logr"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	consulv1alpha1 "github.com/hashicorp/consul-k8s/api/v1alpha1"
+)
+
+// TerminatingGatewayController is the controller for TerminatingGateway resources.
+type TerminatingGatewayController struct {
+	client.Client
+	Log                   logr.Logger
+	Scheme                *runtime.Scheme
+	ConfigEntryController *ConfigEntryController
+}
+
+// +kubebuilder:rbac:groups=consul.hashicorp.com,resources=terminatinggateways,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups=consul.hashicorp.com,resources=terminatinggateways/status,verbs=get;update;patch
+
+func (r *TerminatingGatewayController) Reconcile(req ctrl.Request) (ctrl.Result, error) {
+	return r.ConfigEntryController.ReconcileEntry(r, req, &consulv1alpha1.TerminatingGateway{})
+}
+
+func (r *TerminatingGatewayController) Logger(name types.NamespacedName) logr.Logger {
+	return r.Log.WithValues("request", name)
+}
+
+func (r *TerminatingGatewayController) UpdateStatus(ctx context.Context, obj runtime.Object, opts ...client.UpdateOption) error {
+	return r.Status().Update(ctx, obj, opts...)
+}
+
+func (r *TerminatingGatewayController) SetupWithManager(mgr ctrl.Manager) error {
+	return ctrl.NewControllerManagedBy(mgr).
+		For(&consulv1alpha1.TerminatingGateway{}).
+		Complete(r)
+}